+++ mecon/app/shiny_app.py
@@ -70,7 +70,8 @@ page_title = ui.HTML(f"<big><big><big>mEcon</big></big></big><sub><small><u><i>v
 dataset_label = ui.tooltip(ui.HTML(f"<sub><small>Selected dataset: {get_working_dataset().name}</small></sub>"), f"Dataset directory: {config.DEFAULT_DATASETS_DIR_PATH}")
 navbar = ui.navset_pill(
     ui.nav_control(ui.tags.a("Main page", href=f"http://127.0.0.1:8000/")),
-    # ui.nav_control(ui.tags.a("Reports", href=f"http://127.0.0.1:8001/reports/")),
+    ui.nav_control(ui.tags.a("Datasets", href=f"http://127.0.0.1:8000/datasets")),
+    ui.nav_control(ui.tags.a("Data Flow", href=f"http://127.0.0.1:8000/data")),
     ui.nav_control(ui.tags.a("Edit data", href=f"http://127.0.0.1:8002/edit_data/")),
     # ui.nav_control(ui.tags.a("Monitoring", href=f"http://127.0.0.1:8003/")),
     # ui.nav_control(ui.input_dark_mode(id="light_mode")),
